# packages/contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Smart Travel Assistant — Public API
  version: 1.0.0
  summary: The only HTTP surface a browser is allowed to talk to.
  description: |
    Public trust boundary for the Smart Travel Assistant (module 02). The browser reaches
    `web` and `api` and nothing else: internal services are not routable from outside the
    Docker network, so every capability a user has is expressed here.

    Rules that hold for every operation in this document:

    - Authentication is an OIDC access token issued by Keycloak, verified against JWKS
      (issuer, audience, signature, expiry, not-before). `/health/*` is the only exception.
    - Ownership is resolved from the token, never from a body field. A request that names
      another user's resource gets 404, not 403, so that ids cannot be probed.
    - Times are ISO-8601 with an explicit offset. Coordinates are GeoJSON `[longitude, latitude]`.
    - Every mutating request accepts `Idempotency-Key`. Replaying a key with the same payload
      returns the original result; replaying it with a different payload is `IDEMPOTENCY_CONFLICT`.
    - Every response carries `meta.request_id`, `meta.contract_version` and `meta.generated_at`,
      and lists any dependency that degraded while producing it.
    - Anything safety-relevant carries its source, freshness and quality. When a provider or
      credential is missing the response says so explicitly; it never substitutes a plausible value.
...

...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/runs/{request_id}:
    parameters:
      - $ref: '#/components/parameters/RequestIdPath'
    get:
      tags: [assessments]
      operationId: getRun
      summary: Poll an assessment run
      description: The fallback for clients that cannot hold an SSE connection open.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ContractVersion'
      responses:
        '200':
          description: Current run state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStateResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [assessments]
      operationId: cancelRun
      summary: Cancel an assessment run
      description: |
        Propagates cancellation to the agent and downstream services. The audit trail of what
        already ran is kept; only the work still outstanding is abandoned.
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/ContractVersion'
      responses:
        '200':
          description: Cancellation recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStateResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: The run already reached a terminal state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/v1/runs/{request_id}/events:
    parameters:
      - $ref: '#/components/parameters/RequestIdPath'
    get:
      tags: [assessments]
      operationId: streamRunEvents
      summary: Follow an assessment run over SSE
      description: |
        Server-sent events for one run. Ownership is checked before the stream opens.

        Each event has a monotonic `id`; a client that reconnects with `Last-Event-ID` resumes
        after the last event it saw, so a dropped connection never loses the terminal event.
        `heartbeat` keeps intermediaries from closing an idle stream. The server closes the
        stream after `run.completed`, `run.failed` or cancellation.

        Event names and payloads:

        | event | payload |
        | --- | --- |
        | `run.accepted` | `SseRunAccepted` |
        | `run.progress` | `SseRunProgress` |
        | `run.needs_input` | `SseRunNeedsInput` |
        | `run.degraded` | `SseRunDegraded` |
        | `run.completed` | `SseRunCompleted` |
        | `run.failed` | `SseRunFailed` |
        | `heartbeat` | `SseHeartbeat` |

        No event carries a raw provider body, a prompt, chain-of-thought, an access token or
        personal data. The number of concurrent streams per user and per IP is capped.
      parameters:
        - $ref: '#/components/parameters/ContractVersion'
        - name: Last-Event-ID
          in: header
          required: false
          description: Resume point from a previous connection.
          schema:
            type: string
            maxLength: 64
      responses:
        '200':
          description: |
            An open `text/event-stream`. Sent with `Cache-Control: no-cache` and
            `X-Accel-Buffering: no` so no proxy buffers the progress events.
          headers:
            Cache-Control:
              schema: { type: string }
            X-Accel-Buffering:
              schema: { type: string }
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE frames; see the payload schemas listed above.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429':
          description: Too many concurrent streams for this user or IP.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token issued by the project's Keycloak realm. Verified against the realm's
        JWKS: issuer, audience, signature, expiry and not-before are all checked, and an
        unverifiable token is rejected with 401 — never decoded unsigned as a fallback.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: |
        Client-supplied correlation handle. The server generates one when absent and always
        echoes the value it used in `meta.request_id`.
      schema:
        $ref: '../jsonschema/common/primitives.schema.json#/$defs/Uuid'
    ContractVersion:
      name: X-Contract-Version
      in: header
      required: false
      description: Major contract version the client was built against. Currently `1`.
      schema:
        type: string
        pattern: '^[0-9]+$'
...
    RequestIdPath:
      name: request_id
      in: path
      required: true
      schema:
        $ref: '../jsonschema/common/primitives.schema.json#/$defs/Uuid'
...
  schemas:
    ResponseMeta:
      $ref: '../jsonschema/common/envelope.schema.json#/$defs/ResponseMeta'
...
    ErrorResponse:
      $ref: '../jsonschema/common/envelope.schema.json#/$defs/ErrorResponse'
...
    RunState:
      $ref: '../jsonschema/common/run-state.schema.json'
...
    RunStateResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/RunState'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
  responses:
...
    Unauthorized:
      description: No token, or a token that failed verification.
      headers:
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: |
        No such resource for this caller. Resources owned by another user are reported as not
        found rather than forbidden, so that ids cannot be probed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...